Modify some of the files to upgrade to new versions of Tensorflow and AI Platform.
The legacy Chicago Taxi project utilizes deprecated runtimes for AI Platform (which is in and of itself a deprecated tool). This commit starts making the necessary changes in order to upgrade the project to newer versions of the AI Platform runtime, so that we can re-use this project in order to complete our renewal of the Google Cloud ML Specialization. This is a work-in-progress.